LayoutTests/gamepad/gamepad-polling-access-expected.txt

PASS navigator.webkitGamepads.length is 4
PASS navigator.webkitGamepads[0] is undefined.
PASS navigator.webkitGamepads[1] is undefined.
PASS navigator.webkitGamepads[2] is undefined.
PASS navigator.webkitGamepads[3] is undefined.
PASS navigator.webkitGamepads[0] is non-null.
PASS navigator.webkitGamepads[0] is undefined.
PASS navigator.webkitGamepads[0].id is 'MockStick 3000'
PASS navigator.webkitGamepads[0].buttons.length is 2
PASS navigator.webkitGamepads[0].axes.length is 2
PASS navigator.webkitGamepads[0].buttons[0] is 1.0
PASS navigator.webkitGamepads[0].buttons[1] is 0.0
PASS navigator.webkitGamepads[0].axes.length is 2
PASS navigator.webkitGamepads[0].axes[0] is 0.5
PASS navigator.webkitGamepads[0].axes[1] is -1.0
PASS successfullyParsed is true

TEST COMPLETE
Typical polling access to gamepads contents.
